+++ package.json
@@ -27,6 +27,7 @@
     "extend": "^2.0.1",
     "inflected": "^1.1.6",
     "lodash": "^3.5.0",
+    "merge-descriptors": "^1.0.0",
     "request": "^2.53.0",
     "sax": "^0.6.1",
     "xml2js": "^0.4.6"
